Refactor/cleanup. Velocity multiplier
