+++ crest/Assets/Crest/Crest/Scripts/FX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de16e131af1d40248b315341c6677028
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
